Create header and add title
Use the largest heading size <h1> to create a title on the page
with the user's name. Also, fill in the title between <title></title>.
This doesn't have an effect on the display, but if you check the tab
in your browser, you will see that is now shows your title.